% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{both_ends}
\alias{both_ends}
\title{Target both algebraic ends.}
\usage{
both_ends(k_low, k_high)
}
\arguments{
\item{k_low}{Number of values to select from the smallest algebraic end.}

\item{k_high}{Number of values to select from the largest algebraic end.}
}
\value{
An \code{eigencore_target} descriptor selecting both algebraic ends
(ARPACK \code{BE}).
}
\description{
Target both algebraic ends.
}
