% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean vector space descriptor.}
\usage{
euclidean(dim, dtype = "double")
}
\arguments{
\item{dim}{Dimension of the space.}

\item{dtype}{Scalar type (currently only \code{"double"}).}
}
\value{
An \code{eigencore_space} descriptor for the Euclidean space
\code{R^dim} or \code{C^dim}.
}
\description{
Euclidean vector space descriptor.
}
