% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{general}
\alias{general}
\title{General operator structure descriptor.}
\usage{
general()
}
\value{
An \code{eigencore_structure} descriptor for general operators.
}
\description{
General operator structure descriptor.
}
