% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{hermitian}
\alias{hermitian}
\title{Hermitian/symmetric operator structure descriptor.}
\usage{
hermitian()
}
\value{
An \code{eigencore_structure} descriptor marking an operator as
Hermitian / symmetric.
}
\description{
Hermitian/symmetric operator structure descriptor.
}
