% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{largest}
\alias{largest}
\title{Target the largest algebraic values.}
\usage{
largest()
}
\value{
An \code{eigencore_target} descriptor selecting the largest algebraic
eigenvalues or singular values.
}
\description{
Target the largest algebraic values.
}
