% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{largest_imaginary}
\alias{largest_imaginary}
\title{Target the largest imaginary part.}
\usage{
largest_imaginary()
}
\value{
An \code{eigencore_target} descriptor selecting values by largest
\code{Im(x)} (ARPACK \code{LI}).
}
\description{
Target the largest imaginary part.
}
