% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{largest_real}
\alias{largest_real}
\title{Target the largest real part.}
\usage{
largest_real()
}
\value{
An \code{eigencore_target} descriptor selecting values by largest
\code{Re(x)} (ARPACK \code{LR}).
}
\description{
Target the largest real part.
}
