% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonalization.R
\name{mgs2}
\alias{mgs2}
\title{Modified Gram-Schmidt with two passes.}
\usage{
mgs2(X, against = NULL, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{X}{Numeric matrix input.}

\item{against}{Optional basis to orthogonalize against.}

\item{tol}{Numerical tolerance.}

}
\description{
Modified Gram-Schmidt with two passes.
}
