% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator_algebra.R
\name{operator_scale}
\alias{operator_scale}
\title{Multiply an operator by a scalar.}
\usage{
operator_scale(A, scalar, name = NULL)
}
\arguments{
\item{A}{Matrix or operator input.}

\item{scalar}{Scalar multiplier.}

\item{name}{Optional operator name.}

}
\description{
Multiply an operator by a scalar.
}
