% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator_algebra.R
\name{operator_sum}
\alias{operator_sum}
\title{Sum compatible operators.}
\usage{
operator_sum(..., name = NULL)
}
\arguments{
\item{...}{Additional arguments passed to methods.}

\item{name}{Optional operator name.}

}
\description{
Sum compatible operators.
}
