% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonalization.R
\name{orthogonality_loss}
\alias{orthogonality_loss}
\title{Measure orthogonality loss.}
\usage{
orthogonality_loss(Q, B = NULL)
}
\arguments{
\item{Q}{Basis matrix.}

\item{B}{Metric, operator, or basis input.}

}
\description{
Measure orthogonality loss.
}
\keyword{internal}
