% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{psd_identity}
\alias{psd_identity}
\title{Construct a certified identity PSD factor}
\usage{
psd_identity(dim, policy = psd_policy())
}
\arguments{
\item{dim}{Non-negative whole dimension.}

\item{policy}{A PSD policy from \code{\link[=psd_policy]{psd_policy()}}.}
}
\value{
An \code{eigencore_psd_factor}.
}
\description{
Construct a certified identity PSD factor
}
