% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{psd_operator}
\alias{psd_operator}
\title{Expose a certified PSD action as an eigencore operator}
\usage{
psd_operator(
  x,
  action = c("form", "sqrt", "inverse_sqrt", "pseudoinverse", "image_projector",
    "null_projector")
)
}
\arguments{
\item{x}{A certified PSD factor.}

\item{action}{Certified square action.}
}
\value{
An \code{eigencore_operator}.
}
\description{
Expose a certified PSD action as an eigencore operator
}
