% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{psd_rank}
\alias{psd_rank}
\title{Return certified tolerance-relative PSD rank}
\usage{
psd_rank(x)
}
\arguments{
\item{x}{A certified PSD factor.}
}
\value{
One integer numerical rank.
}
\description{
Return certified tolerance-relative PSD rank
}
