% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/psd.R
\name{psd_reduced_operator}
\alias{psd_reduced_operator}
\title{Construct the Euclidean operator induced on a certified PSD image}
\usage{
psd_reduced_operator(x, A)
}
\arguments{
\item{x}{A certified PSD factor.}

\item{A}{A compatible real-double eigencore operator or matrix.}
}
\value{
An \code{eigencore_operator} for \verb{R+^T A R+}.
}
\description{
Construct the Euclidean operator induced on a certified PSD image
}
