% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certification.R
\name{right_vectors}
\alias{right_vectors}
\title{Extract right singular vectors.}
\usage{
right_vectors(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments passed to methods.}

}
\description{
Extract right singular vectors.
}
