% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator_algebra.R
\name{scale_cols}
\alias{scale_cols}
\title{Scale operator columns.}
\usage{
scale_cols(A, weights, name = NULL)
}
\arguments{
\item{A}{Matrix or operator input.}

\item{weights}{Scaling weights.}

\item{name}{Optional operator name.}

}
\description{
Scale operator columns.
}
