% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{smallest}
\alias{smallest}
\title{Target the smallest algebraic values.}
\usage{
smallest()
}
\value{
An \code{eigencore_target} descriptor selecting the smallest algebraic
eigenvalues or singular values.
}
\description{
Target the smallest algebraic values.
}
