% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{smallest_imaginary}
\alias{smallest_imaginary}
\title{Target the smallest imaginary part.}
\usage{
smallest_imaginary()
}
\value{
An \code{eigencore_target} descriptor selecting values by smallest
\code{Im(x)} (ARPACK \code{SI}).
}
\description{
Target the smallest imaginary part.
}
