% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{smallest_real}
\alias{smallest_real}
\title{Target the smallest real part.}
\usage{
smallest_real()
}
\value{
An \code{eigencore_target} descriptor selecting values by smallest
\code{Re(x)} (ARPACK \code{SR}).
}
\description{
Target the smallest real part.
}
