% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{adjoint}
\alias{adjoint}
\title{Return the adjoint operator.}
\usage{
adjoint(x, ...)
}
\arguments{
\item{x}{Operator-like object.}

\item{...}{Additional arguments passed to methods.}
}
\value{
An \code{eigencore_operator} representing the adjoint map.
}
\description{
Return the adjoint operator.
}
