% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{as_operator}
\alias{as_operator}
\title{Convert an object to an eigencore operator.}
\usage{
as_operator(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments passed to methods.}

}
\description{
Convert an object to an eigencore operator.
}
