% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{auto}
\alias{auto}
\title{Automatic solver choice.}
\usage{
auto()
}
\value{
An \code{eigencore_method} descriptor that lets the planner choose a
solver based on problem structure.
}
\description{
Automatic solver choice.
}
