% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonalization.R
\name{b_orthogonalize}
\alias{b_orthogonalize}
\title{Orthogonalize columns in the B-inner product.}
\usage{
b_orthogonalize(X, B, against = NULL, tol = sqrt(.Machine$double.eps))
}
\arguments{
\item{X}{Numeric matrix input.}

\item{B}{Metric, operator, or basis input.}

\item{against}{Optional basis to orthogonalize against.}

\item{tol}{Numerical tolerance.}

}
\description{
Orthogonalize columns in the B-inner product.
}
\keyword{internal}
