% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certification.R
\name{backward_error}
\alias{backward_error}
\title{Extract backward-error diagnostics.}
\usage{
backward_error(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments passed to methods.}

}
\description{
Extract backward-error diagnostics.
}
