% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator_algebra.R
\name{center}
\alias{center}
\title{Center an operator by rows or columns.}
\usage{
center(
  A,
  rows = FALSE,
  columns = TRUE,
  row_means = NULL,
  col_means = NULL,
  name = NULL
)
}
\arguments{
\item{A}{Matrix or operator input.}

\item{rows}{Whether to operate on rows.}

\item{columns}{Whether to operate on columns.}

\item{row_means}{Optional row means.}

\item{col_means}{Optional column means.}

\item{name}{Optional operator name.}

}
\description{
Center an operator by rows or columns.
}
