% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certification.R
\name{certificate}
\alias{certificate}
\title{Extract a result certificate.}
\usage{
certificate(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments passed to methods.}

}
\description{
Extract a result certificate.
}
