% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certification.R
\name{certificate}
\alias{certificate}
\title{Extract a result certificate.}
\usage{
certificate(x, ...)
}
\arguments{
\item{x}{An eigencore result object.}

\item{...}{Reserved for future methods.}
}
\description{
Extract a result certificate.
}
\examples{
fit <- eig_partial(diag(c(3, 2, 1)), k = 1, target = largest())
cert <- certificate(fit)
cert$passed
cert$max_residual
}
