% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator_algebra.R
\name{compose}
\alias{compose}
\title{Compose two operators.}
\usage{
compose(A, B, name = NULL)
}
\arguments{
\item{A}{Matrix or operator input.}

\item{B}{Metric, operator, or basis input.}

\item{name}{Optional operator name.}

}
\description{
Compose two operators.
}
