% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator_algebra.R
\name{crossprod_operator}
\alias{crossprod_operator}
\title{Create A^* A as an operator.}
\usage{
crossprod_operator(A, name = NULL)
}
\arguments{
\item{A}{Matrix or operator input.}

\item{name}{Optional operator name.}

}
\description{
Create A^* A as an operator.
}
