% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certification.R
\name{diagnostics}
\alias{diagnostics}
\title{Extract diagnostics.}
\usage{
diagnostics(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments passed to methods.}

}
\description{
Extract diagnostics.
}
