% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\name{eigen_problem}
\alias{eigen_problem}
\title{Define an eigenproblem.}
\usage{
eigen_problem(
  A,
  metric = NULL,
  structure = NULL,
  target = largest(),
  transform = NULL
)
}
\arguments{
\item{A}{Matrix or operator input.}

\item{metric}{Optional metric operator.}

\item{structure}{Operator structure descriptor.}

\item{target}{Spectral target descriptor.}

\item{transform}{Optional spectral transform.}

}
\description{
Define an eigenproblem.
}
