% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigencore-package.R
\docType{package}
\name{eigencore-package}
\alias{eigencore}
\alias{eigencore-package}
\title{eigencore}
\description{
Foundations for certified partial spectral computation in R.
}
\author{
\strong{Maintainer}: Bradley Buchsbaum \email{brad.buchsbaum@gmail.com}

}
\keyword{internal}
