% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{eigs}
\alias{eigs}
\title{RSpectra-compatible eigen shim.}
\usage{
eigs(A, k, which = "LM", opts = list(), ...)
}
\arguments{
\item{A}{Matrix or operator input.}

\item{k}{Number of eigenpairs or singular triplets requested.}

\item{which}{RSpectra-style target selector.}

\item{opts}{Compatibility options.}

\item{...}{Additional arguments passed to methods.}

}
\description{
RSpectra-compatible eigen shim.
}
