% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{eigs_sym}
\alias{eigs_sym}
\title{RSpectra-compatible symmetric eigen shim.}
\usage{
eigs_sym(A, k, which = "LA", opts = list(), ...)
}
\arguments{
\item{A}{Matrix or operator input.}

\item{k}{Number of eigenpairs or singular triplets requested.}

\item{which}{RSpectra-style target selector.}

\item{opts}{Compatibility options.}

\item{...}{Additional arguments passed to methods.}

}
\description{
RSpectra-compatible symmetric eigen shim.
}
