% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classes.R
\name{golub_kahan}
\alias{golub_kahan}
\title{Golub-Kahan bidiagonalization method descriptor.}
\usage{
golub_kahan(max_subspace = NULL, reorthogonalize = TRUE)
}
\arguments{
\item{max_subspace}{Optional maximum Krylov subspace size.}

\item{reorthogonalize}{Whether to apply full reorthogonalization.}
}
\value{
An \code{eigencore_method} descriptor selecting Golub-Kahan
bidiagonalization.
}
\description{
Golub-Kahan bidiagonalization method descriptor.
}
