% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certification.R
\name{left_vectors}
\alias{left_vectors}
\title{Extract left singular vectors.}
\usage{
left_vectors(x, ...)
}
\arguments{
\item{x}{An eigencore SVD result object.}

\item{...}{Reserved for future methods.}
}
\description{
Extract left singular vectors.
}
