% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator.R
\name{linear_operator}
\alias{linear_operator}
\title{Create a block-native linear operator.}
\usage{
linear_operator(
  dim,
  apply,
  apply_adjoint = NULL,
  dtype = "double",
  structure = general(),
  name = NULL,
  metadata = list()
)
}
\arguments{
\item{dim}{Operator dimensions.}

\item{apply}{Block apply callback.}

\item{apply_adjoint}{Adjoint block apply callback.}

\item{dtype}{Scalar dtype.}

\item{structure}{Operator structure descriptor.}

\item{name}{Optional operator name.}

\item{metadata}{Optional operator metadata.}

}
\description{
Create a block-native linear operator.
}
