% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/orthogonalization.R
\name{orthogonality_loss}
\alias{orthogonality_loss}
\title{Measure orthogonality loss.}
\usage{
orthogonality_loss(Q, B = NULL)
}
\arguments{
\item{Q}{Matrix whose columns are expected to be orthonormal.}

\item{B}{Optional metric matrix for B-orthogonality.}
}
\description{
Measure orthogonality loss.
}
