% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\name{plan_solver}
\alias{plan_solver}
\title{Plan a solver for a problem.}
\usage{
plan_solver(problem, ...)
}
\arguments{
\item{problem}{Documented argument.}

\item{...}{Additional arguments passed to methods.}

}
\description{
Plan a solver for a problem.
}
