% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\name{plan_solver}
\alias{plan_solver}
\title{Plan a solver for a problem.}
\usage{
plan_solver(problem, ...)
}
\arguments{
\item{problem}{Eigencore eigen or SVD problem object.}

\item{...}{Additional planning arguments passed to methods.}
}
\description{
Plan a solver for a problem.
}
\examples{
A <- diag(c(4, 3, 2, 1))
plan <- plan_solver(eigen_problem(A), k = 2)
plan$method
plan$reasons
}
