% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certification.R
\name{residuals}
\alias{residuals}
\title{Extract residual diagnostics.}
\usage{
residuals(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments passed to methods.}

}
\description{
Extract residual diagnostics.
}
