% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator_algebra.R
\name{scale_rows}
\alias{scale_rows}
\title{Scale operator rows.}
\usage{
scale_rows(A, weights, name = NULL)
}
\arguments{
\item{A}{Matrix or operator input.}

\item{weights}{Scaling weights.}

\item{name}{Optional operator name.}

}
\description{
Scale operator rows.
}
