% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/problem.R
\name{svd_problem}
\alias{svd_problem}
\title{Define an SVD problem.}
\usage{
svd_problem(A, domain = NULL, codomain = NULL, target = largest())
}
\arguments{
\item{A}{Matrix or operator input.}

\item{domain}{Domain space.}

\item{codomain}{Codomain space.}

\item{target}{Spectral target descriptor.}

}
\description{
Define an SVD problem.
}
