% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compatibility.R
\name{svds}
\alias{svds}
\title{RSpectra-compatible SVD shim.}
\usage{
svds(A, k, nu = k, nv = k, opts = list(), ...)
}
\arguments{
\item{A}{Matrix or operator input.}

\item{k}{Number of eigenpairs or singular triplets requested.}

\item{nu}{Number of left singular vectors.}

\item{nv}{Number of right singular vectors.}

\item{opts}{Compatibility options.}

\item{...}{Additional arguments passed to methods.}

}
\description{
RSpectra-compatible SVD shim.
}
