% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operator_algebra.R
\name{symmetric_operator}
\alias{symmetric_operator}
\title{Mark an operator as symmetric/Hermitian.}
\usage{
symmetric_operator(A, validate = TRUE, tol = 1e-10)
}
\arguments{
\item{A}{Matrix or operator input.}

\item{validate}{Whether to validate assumptions.}

\item{tol}{Numerical tolerance.}

}
\description{
Mark an operator as symmetric/Hermitian.
}
