% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certification.R
\name{values}
\alias{values}
\title{Extract computed values.}
\usage{
values(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments passed to methods.}

}
\description{
Extract computed values.
}
