% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certification.R
\name{vectors}
\alias{vectors}
\title{Extract eigenvectors.}
\usage{
vectors(x, ...)
}
\arguments{
\item{x}{An eigencore eigen result object.}

\item{...}{Reserved for future methods.}
}
\description{
Extract eigenvectors.
}
\examples{
fit <- eig_partial(diag(c(3, 2, 1)), k = 2, target = largest())
dim(vectors(fit))
}
