% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/certification.R
\name{vectors}
\alias{vectors}
\title{Extract eigenvectors.}
\usage{
vectors(x, ...)
}
\arguments{
\item{x}{Object.}

\item{...}{Additional arguments passed to methods.}

}
\description{
Extract eigenvectors.
}
